requirements/_base.txt
amqp==5.1.1
asgiref==3.5.2
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.2
Django==4.0.6
django-browser-reload==1.6.0
django-tailwind==3.3.0
fatsecret==0.4.0
rauth
iniconfig==1.1.1
kombu==5.2.4
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.5
crispy-tailwind
django-braces
